refine lenet
