+++ pyspark/bigdl/examples/lenet/lenet.py
@@ -15,10 +15,8 @@
 #
 
 from optparse import OptionParser
-from bigdl.dataset import mnist
-from bigdl.dataset.transformer import *
+from bigdl.models.lenet.utils import *
 from bigdl.nn.keras.layer import *
-from bigdl.nn.layer import Model
 from bigdl.nn.criterion import *
 from bigdl.optim.optimizer import *
 from bigdl.util.common import *
@@ -38,28 +36,9 @@ def build_model(class_num):
     return model
 
 
-def get_mnist(sc, data_type="train", location="/tmp/mnist"):
-    """
-    Get and normalize the mnist data. We would download it automatically
-    if the data doesn't present at the specific location.
-
-    :param sc: SparkContext
-    :param data_type: training data or testing data
-    :param location: Location storing the mnist
-    :return: A RDD of (features: Ndarray, label: Ndarray)
-    """
-    (images, labels) = mnist.read_data_sets(location, data_type)
-    images = sc.parallelize(images)
-    labels = sc.parallelize(labels + 1)  # Target start from 1 in BigDL
-    record = images.zip(labels)
-    return record
-
-
 if __name__ == "__main__":
     parser = OptionParser()
-    parser.add_option("-a", "--action", dest="action", default="train")
     parser.add_option("-b", "--batchSize", type=int, dest="batchSize", default="128")
-    parser.add_option("-o", "--modelPath", dest="modelPath", default="/tmp/lenet5/model.470")
     parser.add_option("-c", "--checkpointPath", dest="checkpointPath", default="/tmp/lenet5")
     parser.add_option("-t", "--endTriggerType", dest="endTriggerType", default="epoch")
     parser.add_option("-n", "--endTriggerNum", type=int, dest="endTriggerNum", default="20")
@@ -72,43 +51,22 @@ if __name__ == "__main__":
     show_bigdl_info_logs()
     init_engine()
 
-    if options.action == "train":
-        def get_end_trigger():
-            if options.endTriggerType.lower() == "epoch":
-                return MaxEpoch(options.endTriggerNum)
-            else:
-                return MaxIteration(options.endTriggerNum)
+    (train_data, test_data) = process_mnist_rdd(sc, options)
 
-        train_data = get_mnist(sc, "train", options.dataPath)\
-            .map(lambda rec_tuple: (normalizer(rec_tuple[0], mnist.TRAIN_MEAN, mnist.TRAIN_STD),
-                               rec_tuple[1]))\
-            .map(lambda t: Sample.from_ndarray(t[0], t[1]))
-        test_data = get_mnist(sc, "test", options.dataPath)\
-            .map(lambda rec_tuple: (normalizer(rec_tuple[0], mnist.TEST_MEAN, mnist.TEST_STD),
-                               rec_tuple[1]))\
-            .map(lambda t: Sample.from_ndarray(t[0], t[1]))
-        optimizer = Optimizer(
-            model=build_model(10),
-            training_rdd=train_data,
-            criterion=ClassNLLCriterion(logProbAsInput=False),
-            optim_method=SGD(learningrate=0.01, learningrate_decay=0.0002),
-            end_trigger=get_end_trigger(),
-            batch_size=options.batchSize)
-        optimizer.set_validation(
-            batch_size=options.batchSize,
-            val_rdd=test_data,
-            trigger=EveryEpoch(),
-            val_method=[Top1Accuracy()]
-        )
-        optimizer.set_checkpoint(EveryEpoch(), options.checkpointPath)
-        trained_model = optimizer.optimize()
-        parameters = trained_model.parameters()
-    elif options.action == "test":
-        # Load a pre-trained model and then validate it through top1 accuracy.
-        test_data = get_mnist(sc, "test").map(
-            normalizer(mnist.TEST_MEAN, mnist.TEST_STD))
-        model = Model.load(options.modelPath)
-        results = model.evaluate(test_data, options.batchSize, [Top1Accuracy()])
-        for result in results:
-            print(result)
+    optimizer = Optimizer(
+        model=build_model(10),
+        training_rdd=train_data,
+        criterion=ClassNLLCriterion(logProbAsInput=False),
+        optim_method=SGD(learningrate=0.01, learningrate_decay=0.0002),
+        end_trigger=get_end_trigger(options),
+        batch_size=options.batchSize)
+    optimizer.set_validation(
+        batch_size=options.batchSize,
+        val_rdd=test_data,
+        trigger=EveryEpoch(),
+        val_method=[Top1Accuracy()]
+    )
+    optimizer.set_checkpoint(EveryEpoch(), options.checkpointPath)
+    trained_model = optimizer.optimize()
+    parameters = trained_model.parameters()
     sc.stop()
